feat: support background color and deduplicate format logic
