+++ popup.html
@@ -16,6 +16,6 @@
   <footer>
     <p>A free tool from <a href="https://www.linkedin.com/in/adenning2/" target="_blank">Adam Denning</a></p>
   </footer>
-  <script src="popup.js"></script>
+  <script type="module" src="popup.js"></script>
 </body>
 </html>
